+++ solutions/gcd/main.cpp
@@ -1,21 +1,21 @@
 #include "main.h"
 
 // using euclid algorithm
-long long gcd(long long a, long long b) {
+int gcd(int a, int b) {
     if (b == 0) return a;
     return gcd(b, a % b);
 }
 
 // using extended euclid algorithm
-long long gcd_extend(long long a, long long b, long long *x, long long *y) {
+int gcd_extend(int a, int b, int *x, int *y) {
     if (a == 0) {
         *x = 0;
         *y = 1;
         return b;
     }
 
-    long long x1, y1;
-    long long gcd = gcd_extend(b % a, a, &x1, &y1);
+    int x1, y1;
+    int gcd = gcd_extend(b % a, a, &x1, &y1);
 
     *x = y1 - (b / a) * x1;
     *y = x1;
